+++ .nyc_output/processinfo/index.json
@@ -1,1 +1,1 @@
-{"processes":{"93e1952c-2649-4e86-a332-d8cde2e8e11b":{"parent":null,"children":[]}},"files":{"/home/pascal/Documents/teamwork/api/index.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/Auth/signin.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/helpers/validateSignIn.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/models/users.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/Auth/signup.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/helpers/validateUser.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/controllers/postArticle.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/helpers/validateArticle.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/models/article.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/controllers/editArticle.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/controllers/deleteArticle.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/controllers/comentArticle.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/helpers/validateComment.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/controllers/viewArticles.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/controllers/specificArticle.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/models/comment.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"],"/home/pascal/Documents/teamwork/api/middleware/auth.js":["93e1952c-2649-4e86-a332-d8cde2e8e11b"]},"externalIds":{}}
+{"processes":{"edc180b1-4f91-4da1-8161-31e623216364":{"parent":null,"children":[]}},"files":{"/home/pascal/Documents/teamwork/api/index.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/signin.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/helpers/validateSignIn.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/models/users.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/signup.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/helpers/validateUser.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/postArticle.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/helpers/validateArticle.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/models/article.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/editArticle.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/deleteArticle.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/comentArticle.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/helpers/validateComment.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/viewArticles.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/controllers/specificArticle.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/models/comment.js":["edc180b1-4f91-4da1-8161-31e623216364"],"/home/pascal/Documents/teamwork/api/middleware/auth.js":["edc180b1-4f91-4da1-8161-31e623216364"]},"externalIds":{}}
