+++ api/index.js
@@ -1,8 +1,8 @@
 /* eslint-disable no-console */
 import express from 'express';
 import winston from 'winston';
-import signin from './Auth/signin';
-import signup from './Auth/signup';
+import signin from './controllers/signin';
+import signup from './controllers/signup';
 import articlePost from './controllers/postArticle';
 import editArticle from './controllers/editArticle';
 import deleteArticle from './controllers/deleteArticle';
